2020 7月29日

+++ src/main/dabao/com/ecologicalRanch/project/entity/Field.java
@@ -14,7 +14,7 @@ public class Field {
      * 场地ID
      */
 
-    private Long fieldId;
+    private Integer fieldId;
 
     /**
      * 牲畜场管理员id
